+++ event-sourcing/event-repository/event-repository-jdbc/src/test/java/uk/gov/justice/services/eventsourcing/repository/jdbc/JdbcBasedEventRepositoryTest.java
@@ -198,8 +198,22 @@ public class JdbcBasedEventRepositoryTest {
         assertThat(listOfStreams.get(2).findFirst().get(), is(envelope3));
     }
 
-    private EventStream buildEventStreamFor(final UUID streamId, final Long sequence) {
-        return new EventStream(streamId, sequence, true, TIMESTAMP);
+    @Test
+    public void shouldGetActiveStreamOfStreamIds() throws Exception {
+        final UUID streamId1 = UUID.fromString("4b4e80a0-76f7-476c-b75b-527e38fb251e");
+        final UUID streamId2 = UUID.fromString("4b4e80a0-76f7-476c-b75b-527e38fb252e");
+        final UUID streamId3 = UUID.fromString("4b4e80a0-76f7-476c-b75b-527e38fb253e");
+
+        when(eventStreamJdbcRepository.findActive()).thenReturn(Stream.of(buildEventStreamFor(streamId1, 1L), buildEventStreamFor(streamId2, 2L), buildEventStreamFor(streamId3, 3L)));
+
+        final Stream<UUID> allActiveStreamIds = jdbcBasedEventRepository.getAllActiveStreamIds();
+
+        final List<UUID> streamIds = allActiveStreamIds.collect(toList());
+        assertThat(streamIds, hasSize(3));
+
+        assertThat(streamIds.get(0), is(streamId1));
+        assertThat(streamIds.get(1), is(streamId2));
+        assertThat(streamIds.get(2), is(streamId3));
     }
 
     @Test
@@ -366,6 +380,10 @@ public class JdbcBasedEventRepositoryTest {
     public void shouldStoreEventStream() {
         jdbcBasedEventRepository.createEventStream(STREAM_ID);
         verify(eventStreamJdbcRepository).insert(STREAM_ID);
+    }
+
+    private EventStream buildEventStreamFor(final UUID streamId, final Long sequence) {
+        return new EventStream(streamId, sequence, true, TIMESTAMP);
     }
 
     private Event eventOf(final UUID streamId) {